build-mingw-w64: Fix installing COPYING files on macOS
The install(1) tool on macOS doesn't support the -D option for
creating directories at the same time while installing files.
Therefore, run a separate step for creating directories. Doing that
with "mkdir -p", even if install(1) also could do that with the -d
option (likewise with a separate command).

Also stop specifying the destination file name, when it's equal to
the source files, now that we know that the destination directory
already exists.